requirements_base.txt
Django==1.9.4
django-bootstrap-form==3.2
django-braces==1.8.1
# django-celery==3.1.17
# django-localeurl==2.0.2
# django-rosetta==0.7.10
django-annoying
Pillow
django-wysiwyg-redactor
django-compressor==1.6
